// src/features/sensors/domain/models/DeviceState.ts
export type DeviceState = {
  motor:                'encendido' | 'apagado'
  bomba:                'encendido' | 'apagado'
  sensor_temperatura:   'encendido' | 'apagado'
  sensor_ph:            'encendido' | 'apagado'
  sensor_alcohol:       'encendido' | 'apagado'
  sensor_conductividad: 'encendido' | 'apagado'
  sensor_turbidez:      'encendido' | 'apagado'
}